Fixes in psrspa
This revision increases verbosity of the parts of the code taken out of spa.C.
psrspa won't crash anymore when no or only one pulse is found.
Clean up of code related to finding maximal amplitude in chosen phase range.
Changed the format of the output of maximal amplitude, added the baseline rms.