Major speed up of the code
I was calculating proj_covariance matrix (alfa in the paper) in the
slowest way imaginable. Also, its definition was quite confusing as it
was transposed w.r.t. to the definition in the paper.
The code is pretty fast for nbin=1024 but for profiles with more bins it
was extremely slow, now the slow part of the code is sped up by a factor
of about 8000...