FrequencyAppend: Option to force a new predictor.
This fixes the bug from report #400.

A new option was added to force FrequencyAppend class to always generate
a new predictor spanning full time and frequency range of combined into
and from. I decided to implement this as an option as it doesn't affect
many users and it has adverse effect on psrchive's performance as it
increases the number of tempo2 system calls.

Files modified:
More/General/FrequencyAppend.C
More/General/Pulsar/FrequencyAppend.h